Work out the workers and transformation
In this process I figured out that there in theory would be multiple ways to convert a jsonpath to a mongo query where a aggregation might not always be the one and only option.

The main two methods to do a conversion seem to be; To an aggregation (which allows us to use $geoNear for some nifty lookups) and to a find({ $where }) query. Although the performance of using a find would be an issue to say the least it does make it possible to generate code to completely conform to what ever we want in a query.

Going for the low-hanging fruit flesh out the aggregations first.